commonsys-intf: services: upgrade IDC for backward compatible
The IDC libs in vendor and system images are not compatible due
to diverged changes. E.g., in legacy vendor kGetDisplayTileCount
op code is 50, while in new system images op code 50 stands for
kDummyOpCode. This breaks the backward compatibility. So upgrade
IDC interface in system image to fix the backward compatible
issues for vendor freeze targets.

Change-Id: I170c279f958e33e143677c9888aca462a9d2874f