jasmine_sprout: Correct power value for cluster6
This was done by me as a aprox estimation cause this was missing and was added manually.
Thanks a lot to @jjpprrrr for his explanation on how this works and why this is important

Ref https://github.com/PixelExperience-Devices/device_xiaomi_polaris/commit/31c089a2eaee91f7d283144e9ec7f27546a0a5b0

Power of 2208 MHz = 253 * 978 / 802

where 978 and 802 are cpu power measured by freqbench.

https://github.com/Tashar02/freqbench/blob/master/results/sdm660/fifth/results.csv

Thanks Tashar for doing freqbench